+++ qc-application/src/main/java/com/eager/questioncloud/application/api/post/dto/PostCommentControllerRequest.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.post;
+package com.eager.questioncloud.application.api.post.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
