+++ qc-application/src/main/java/com/eager/questioncloud/application/api/post/dto/PostControllerResponse.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.post;
+package com.eager.questioncloud.application.api.post.dto;
 
 import com.eager.questioncloud.core.domain.post.dto.PostDetail;
 import lombok.AllArgsConstructor;
