+++ qc-application/src/main/java/com/eager/questioncloud/application/api/post/service/PostCommentService.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.application.api.post;
+package com.eager.questioncloud.application.api.post.service;
 
+import com.eager.questioncloud.application.api.post.implement.PostPermissionChecker;
 import com.eager.questioncloud.core.common.PagingInformation;
 import com.eager.questioncloud.core.domain.post.dto.PostCommentDetail;
 import com.eager.questioncloud.core.domain.post.infrastructure.PostCommentRepository;
